Refactor duplicate code DefaultEsClientFactory/DefaultESClientFactory classes
