+++ common/es/src/main/resources-filtered/io/apiman/common/es/util/apiman_cache-settings.json
