+++ common/es/src/main/resources-filtered/io/apiman/common/es/util/apiman_gateway-settings.json
