+++ distro/tomcat8/src/main/resources/overlay/conf/apiman.properties
@@ -33,22 +33,22 @@ apiman-manager.security-context.type=default
 
 # API Manager storage settings.
 apiman-manager.storage.type=es
-apiman-manager.storage.es.protocol=${apiman.es.protocol}
-apiman-manager.storage.es.host=${apiman.es.host}
-apiman-manager.storage.es.port=${apiman.es.port}
-apiman-manager.storage.es.username=${apiman.es.username}
-apiman-manager.storage.es.password=${apiman.es.password}
-apiman-manager.storage.es.timeout=${apiman.es.timeout}
-apiman-manager.storage.es.initialize=true
+apiman-manager.storage.es.client.protocol=${apiman.es.protocol}
+apiman-manager.storage.es.client.host=${apiman.es.host}
+apiman-manager.storage.es.client.port=${apiman.es.port}
+apiman-manager.storage.es.client.username=${apiman.es.username}
+apiman-manager.storage.es.client.password=${apiman.es.password}
+apiman-manager.storage.es.client.timeout=${apiman.es.timeout}
+apiman-manager.storage.es.client.initialize=true
 
 # API Manager metrics settings.
 apiman-manager.metrics.type=es
-apiman-manager.metrics.es.protocol=${apiman.es.protocol}
-apiman-manager.metrics.es.host=${apiman.es.host}
-apiman-manager.metrics.es.port=${apiman.es.port}
-apiman-manager.metrics.es.username=${apiman.es.username}
-apiman-manager.metrics.es.password=${apiman.es.password}
-apiman-manager.metrics.es.timeout=${apiman.es.timeout}
+apiman-manager.metrics.es.client.protocol=${apiman.es.protocol}
+apiman-manager.metrics.es.client.host=${apiman.es.host}
+apiman-manager.metrics.es.client.port=${apiman.es.port}
+apiman-manager.metrics.es.client.username=${apiman.es.username}
+apiman-manager.metrics.es.client.password=${apiman.es.password}
+apiman-manager.metrics.es.client.timeout=${apiman.es.timeout}
 
 # API Manager API Catalog
 apiman-manager.api-catalog.type=io.apiman.manager.api.core.catalog.JsonApiCatalog
