+++ gateway/test/src/test/java/io/apiman/gateway/test/junit/vertx3/EsResetter.java
@@ -15,8 +15,8 @@
  */
 package io.apiman.gateway.test.junit.vertx3;
 
+import io.apiman.common.es.util.DefaultEsClientFactory;
 import io.apiman.gateway.engine.es.AbstractESComponent;
-import io.apiman.gateway.engine.es.DefaultESClientFactory;
 import io.apiman.gateway.engine.es.ESConstants;
 import io.apiman.gateway.platforms.vertx3.common.config.VertxEngineConfig;
 import io.searchbox.client.JestResult;
@@ -48,7 +48,7 @@ public class EsResetter extends AbstractESComponent implements Resetter {
 
             // Important! Or will get cached client that assumes the DB schema has already been created
             // and subtly horrible things will happen, and you'll waste a whole day debugging it! :-)
-            DefaultESClientFactory.clearClientCache();
+            DefaultEsClientFactory.clearClientCache();
 
             getClient().executeAsync(new Delete.Builder(getDefaultIndexName()).build(),
                     new JestResultHandler<JestResult>() {
