--- manager/api/es/src/main/java/io/apiman/manager/api/es/IEsClientFactory.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2016 JBoss Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.apiman.manager.api.es;
-
-import io.searchbox.client.JestClient;
-
-/**
- * A factory used to create elasticsearch clients.
- * @author eric.wittmann@gmail.com
- */
-public interface IEsClientFactory {
-    
-    /**
-     * Called to create the ES client.
-     */
-    public JestClient createClient();
-
-}
