+++ manager/api/war/pom.xml
@@ -90,6 +90,10 @@
       <groupId>${project.groupId}</groupId>
       <artifactId>apiman-manager-api-gateway</artifactId>
     </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>apiman-common-es</artifactId>
+    </dependency>
 
     <!-- Third Party Dependencies -->
     <dependency>
