+++ manager/api/war/src/main/java/io/apiman/manager/api/war/WarCdiFactory.java
@@ -15,6 +15,8 @@
  */
 package io.apiman.manager.api.war;
 
+import io.apiman.common.es.util.DefaultEsClientFactory;
+import io.apiman.common.es.util.IEsClientFactory;
 import io.apiman.common.logging.IApimanDelegateLogger;
 import io.apiman.common.logging.IApimanLogger;
 import io.apiman.common.plugin.Plugin;
@@ -40,10 +42,8 @@ import io.apiman.manager.api.core.logging.ApimanLogger;
 import io.apiman.manager.api.core.logging.JsonLoggerImpl;
 import io.apiman.manager.api.core.logging.StandardLoggerImpl;
 import io.apiman.manager.api.core.noop.NoOpMetricsAccessor;
-import io.apiman.manager.api.es.DefaultEsClientFactory;
 import io.apiman.manager.api.es.ESMetricsAccessor;
 import io.apiman.manager.api.es.EsStorage;
-import io.apiman.manager.api.es.IEsClientFactory;
 import io.apiman.manager.api.jpa.JpaStorage;
 import io.apiman.manager.api.jpa.JpaStorageInitializer;
 import io.apiman.manager.api.security.ISecurityContext;
@@ -233,7 +233,7 @@ public class WarCdiFactory {
                     factoryClass = DefaultEsClientFactory.class.getName();
                 }
                 sStorageESClientFactory = createCustomComponent(IEsClientFactory.class, factoryClass,
-                        config.getStorageESClientFactoryConfig(), pluginRegistry);
+                        null, pluginRegistry);
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
@@ -250,7 +250,7 @@ public class WarCdiFactory {
                     factoryClass = DefaultEsClientFactory.class.getName();
                 }
                 sMetricsESClientFactory = createCustomComponent(IEsClientFactory.class, factoryClass,
-                        config.getMetricsESClientFactoryConfig(), pluginRegistry);
+                        null, pluginRegistry);
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
@@ -261,7 +261,7 @@ public class WarCdiFactory {
     @Produces @ApplicationScoped @Named("storage")
     public static JestClient provideStorageESClient(WarApiManagerConfig config, @Named("storage-factory") IEsClientFactory clientFactory) {
         if ("es".equals(config.getStorageType())) { //$NON-NLS-1$
-            return clientFactory.createClient();
+            return clientFactory.createClient(config.getMetricsESClientFactoryConfig(), null);
         } else {
             return null;
         }
@@ -270,12 +270,12 @@ public class WarCdiFactory {
     @Produces @ApplicationScoped @Named("metrics")
     public static JestClient provideMetricsESClient(WarApiManagerConfig config, @Named("metrics-factory") IEsClientFactory clientFactory) {
         if ("es".equals(config.getMetricsType())) { //$NON-NLS-1$
-            return clientFactory.createClient();
+            return clientFactory.createClient(config.getStorageESClientFactoryConfig(), null);
         } else {
             return null;
         }
     }
-    
+
     /**
      * Initializes the ES storage (if required).
      * @param config
