+++ manager/test/api/src/test/java/io/apiman/manager/test/es/ESMetricsAccessorTest.java
@@ -17,7 +17,7 @@
 package io.apiman.manager.test.es;
 
 import io.apiman.common.es.util.ApimanEmbeddedElastic;
-import io.apiman.gateway.engine.es.DefaultESClientFactory;
+import io.apiman.common.es.util.DefaultEsClientFactory;
 import io.apiman.manager.api.beans.metrics.ClientUsagePerApiBean;
 import io.apiman.manager.api.beans.metrics.HistogramIntervalType;
 import io.apiman.manager.api.beans.metrics.ResponseStatsDataPoint;
@@ -93,7 +93,7 @@ public class ESMetricsAccessorTest {
         client = createJestClient();
         client.execute(new DeleteIndex.Builder("apiman_metrics").build());
         client.execute(new Flush.Builder().force().build());
-        DefaultESClientFactory.clearClientCache();
+        DefaultEsClientFactory.clearClientCache();
         // Because of the delete above, the metrics fields need reinitialising with the index
         // mappings otherwise everything will screw up. See apiman_metrics-settings.json
         client = createJestClient();
@@ -102,7 +102,7 @@ public class ESMetricsAccessorTest {
         loadTestData();
 
         client.execute(new Flush.Builder().force().build());
-        DefaultESClientFactory.clearClientCache();
+        DefaultEsClientFactory.clearClientCache();
     }
 
     private static JestClient createJestClient() {
@@ -111,7 +111,7 @@ public class ESMetricsAccessorTest {
         config.put("client.host", "localhost");
         config.put("client.port", "19250");
         config.put("client.initialize", "true");
-        return new DefaultESClientFactory().createClient(config, "apiman_metrics");
+        return new DefaultEsClientFactory().createClient(config, "apiman_metrics");
     }
 
     private static void loadTestData() throws Exception {
